Dharam  Created an simple web page which lets you to store your details in data base
What changes have you made?
wrote an server which will respond to the client request and accordingly
talks to the data base and store into or fetch data from the data base.
And also created an simple web page which will collect the information from
the user and send all the data to the server which will store that data into
data base.